backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi>=0.100.0
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pydantic>=2.0.0
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv>=1.0.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.6
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn[standard]>=0.20.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1

# Database (PostgreSQL)
psycopg2-binary>=2.9.0 # For synchronous SQLAlchemy or direct use
# asyncpg>=0.27.0 # For asynchronous SQLAlchemy with PostgreSQL
# sqlmodel>=0.0.8 # Optional, if you prefer SQLModel over SQLAlchemy directly

# Langchain & LangGraph
langchain>=0.1.0
langgraph>=0.0.20

# For embeddings (example, choose one or more as needed)
# langchain-openai>=0.0.1 
# langchain-community
# sentence-transformers # For local embeddings

# For vector stores (example, if using pgvector with Langchain)
# psycopg2cffi # Alternative to psycopg2-binary sometimes needed for pgvector
# pgvector # Python client for pgvector extension

# Optional, for older Langchain versions or specific components
# sqlalchemy
